+++ go-basic-godev/errhandling/defer/defer.go
@@ -4,8 +4,9 @@ import (
 	"bufio"
 	"errors"
 	"fmt"
-	"github.com/upupnoah/go-basic-future/go-basic/errhandling/fib"
 	"os"
+
+	"github.com/upupnoah/go-basic-future/go-basic-godev/errhandling/fib"
 )
 
 func tryDefer() {
