+++ go-core-tony/simple-webserver/hello-go/bookstore/cmd/bookstore/main.go
@@ -2,13 +2,14 @@ package main
 
 import (
 	"context"
-	"github.com/upupnoah/go-basic-future/go-core/simple-webserver/hello-go/bookstore/server"
-	"github.com/upupnoah/go-basic-future/go-core/simple-webserver/hello-go/bookstore/store/factory"
 	"log"
 	"os"
 	"os/signal"
 	"syscall"
 	"time"
+
+	"github.com/upupnoah/go-basic-future/go-core-tony/simple-webserver/hello-go/bookstore/server"
+	"github.com/upupnoah/go-basic-future/go-core-tony/simple-webserver/hello-go/bookstore/store/factory"
 )
 
 func main() {
